requirements.txt
pypdf>=6.4.2
python-dotenv>=1.2.1
google-genai>=1.56.0
tiktoken>=0.12.0
